+++ drawlib/physics.h
@@ -1,8 +1,8 @@
 #ifndef _PHYSICS_
 #define _PHYSICS_
 
-int collisionRect(RECTANGLE r1, RECTANGLE r2);
-int collisionScreenRect(RECTANGLE r);
+short int collisionRect(RECTANGLE r1, RECTANGLE r2);
+short int collisionScreenRect(RECTANGLE r);
 
 
 #endif
